+++ R/display.R
@@ -25,9 +25,12 @@
 #'   data frames, `caption` may be a list of table captions, one for each table.
 #' @param font_size For HTML tables, the font size.
 #' @param line_padding For HTML tables, the distance (in pixel) between lines.
+#' @param column_labels Labels of columns for HTML tables. If `NULL`, automatic
+#'   column names are generated. See 'Examples'.
 #' @inheritParams print.parameters_model
 #' @inheritParams insight::format_table
 #' @inheritParams insight::export_table
+#' @inheritParams compare_parameters
 #'
 #' @return If `format = "markdown"`, the return value will be a character
 #'   vector in markdown-table format. If `format = "html"`, an object of
@@ -44,6 +47,20 @@
 #' model <- lm(mpg ~ wt + cyl, data = mtcars)
 #' mp <- model_parameters(model)
 #' display(mp)
+#'
+#' \dontrun{
+#' data(iris)
+#' lm1 <- lm(Sepal.Length ~ Species, data = iris)
+#' lm2 <- lm(Sepal.Length ~ Species + Petal.Length, data = iris)
+#' lm3 <- lm(Sepal.Length ~ Species * Petal.Length, data = iris)
+#' out <- compare_parameters(lm1, lm2, lm3)
+#'
+#' print_html(
+#'   out,
+#'   style = "{coef}{stars}|({ci})",
+#'   column_labels = c("Estimate", "95% CI")
+#' )
+#' }
 #' @export
 display.parameters_model <- function(object,
                                      format = "markdown",
@@ -64,6 +81,8 @@ display.parameters_model <- function(object,
                                      zap_small = FALSE,
                                      font_size = "100%",
                                      line_padding = 4,
+                                     style = NULL,
+                                     column_labels = NULL,
                                      verbose = TRUE,
                                      ...) {
   if (identical(format, "html")) {
@@ -73,7 +92,8 @@ display.parameters_model <- function(object,
       footer = footer, ci_digits = ci_digits, p_digits = p_digits,
       footer_digits = footer_digits, align = align, ci_brackets = ci_brackets,
       show_sigma = show_sigma, show_formula = show_formula, zap_small = zap_small,
-      font_size = font_size, line_padding = line_padding, verbose = verbose, ...
+      font_size = font_size, line_padding = line_padding, style = style,
+      column_labels = column_labels, verbose = verbose, ...
     )
   } else {
     print_md(
@@ -110,11 +130,19 @@ display.compare_parameters <- function(object,
                                        ci_digits = 2,
                                        p_digits = 3,
                                        style = NULL,
+                                       column_labels = NULL,
+                                       ci_brackets = c("(", ")"),
                                        font_size = "100%",
                                        line_padding = 4,
+                                       zap_small = FALSE,
                                        ...) {
   if (identical(format, "html")) {
-    print_html(x = object, digits = digits, ci_digits = ci_digits, p_digits = p_digits, style = style, font_size = font_size, line_padding = line_padding, ...)
+    print_html(
+      x = object, digits = digits, ci_digits = ci_digits, p_digits = p_digits,
+      style = style, column_labels = column_labels, font_size = font_size,
+      line_padding = line_padding, ci_brackets = ci_brackets,
+      zap_small = zap_small, ...
+    )
   } else {
     print_md(x = object, digits = digits, ci_digits = ci_digits, p_digits = p_digits, style = style, ...)
   }